合并type.go和types.go
